Move sleep helper to its own file
